PDMS2_web_old/requirements.txt
Flask==2.2.5
opencv-python>=4.6.0
opencv-contrib-python>=4.6.0
ultralytics==8.0.100
torch>=1.10.0
torchvision>=0.11.0
tensorflow>=2.8.0
keras>=2.8.0
numpy>=1.19.0
scipy>=1.7.0
Pillow>=8.0.0
scikit-image>=0.18.0
pandas>=1.3.0
pathlib2>=2.3.6